Use the config
This commit modifies the create module to rely on the configuration file recently created. This should make it easier to statically alter the initial behaviour of the app. Note that since the config sets the initial page to `about:blank`, there is currently no need for the placeholder index page.