Add mout dependency
This commit adds the MOUT library as a dependency. MOUT provides a number of handy functions that will save time. It is added as a development dependency for now, alongside electron. NPM also created a package-lock which it said should be included in version control, hence it is included in this commit.