Begin looking for config files
This commit adds a basic structure for a config file search. It searches in the root of the repo, and in its containing folder, for a configuration by the name of "electron-config". The first it finds, it loads and merges into the defaults - hence no configuration file need be specified if the default electron configuration is sufficient. A template file is included to demonstrate how directly the electron settings can be specified - though the development flag is kept separate which may need some documentation at some point. Note that the Git ignore file has been modified to exclude any such configuration files from tracking, to make this easier to use.